+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 2026-07-20
+* RAA-7897: Replace S3/targets.json routing lookup with live EPC API call
+* RAA-7897: Add BuildEpcPath.js to construct Endpoint query from NHSD-Target-Identifier header
+* RAA-7897: Add ParseEpcResponse.js to extract receiver address from FHIR Bundle response
+* RAA-7897: Fix HealthcareService.Identifier query parameter casing (capital I)
+
 ## {{ ansible_date_time.date }}
 * Initial commit